Time: 0 ms (100.00%) | Memory: 9.7 MB (33.73%) - LeetSync

+++ 2561-number-of-distinct-averages/number-of-distinct-averages.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int distinctAverages(vector<int>& nums) {
         sort(nums.begin(),nums.end());
-        set<int>st;
+        set<long long >st;
         int i=0 ; 
         int j=nums.size()-1;
         while(i<j){
